backend: extract relations to separate file
