+++ backend/src/db/index.ts
@@ -4,11 +4,19 @@ import env from '../lib/env';
 import * as directory from './schema/directory.schema';
 import * as file from './schema/file.schema';
 import * as message from './schema/message.schema';
+import * as relations from './schema/relations.schema';
 import * as thread from './schema/thread.schema';
 import * as user from './schema/user.schema';
 
 const sqlite = new Database(env.DB_FILE_NAME);
 export const db = drizzle({
   client: sqlite,
-  schema: { ...directory, ...message, ...thread, ...user, ...file }
+  schema: {
+    ...directory,
+    ...message,
+    ...thread,
+    ...user,
+    ...file,
+    ...relations
+  }
 });
